refactor: runCode 버튼 위치 조정
